requirements.txt
openai
pandas
openpyxl
python-dotenv
tqdm
scikit-learn
pydantic
tenacity